move openXDA stats into separate component
