+++ Source/Applications/SystemCenter/wwwroot/Scripts/TSX/SystemCenter/AppHost/NodeDetails.tsx
@@ -25,9 +25,7 @@
 
 import * as React from 'react';
 import { Modal, TabSelector } from '@gpa-gemstone/react-interactive';
-import { Application } from '@gpa-gemstone/application-typings';
 import NodeHealth from './NodeHealth';
-import { ReactIcons } from '@gpa-gemstone/gpa-symbols';
 
 type tabs = 'Connections' | 'Health' | 'Console'
 
@@ -41,100 +39,23 @@ export interface IProps {
     Properties: { Name: string, Value: string }[]
 }
 
-const statStyle: React.CSSProperties = {
-    fontSize: "1em",
-    display: 'inline-block',
-    overflow: 'auto',
-    whiteSpace: 'pre-wrap',
-    wordBreak: 'break-word'
-}
 
 const NodeDetails = (props: IProps) => {
-    const [statInfo, setStatInfo] = React.useState<string>('');
-    const [status, setStatus] = React.useState<Application.Types.Status>('uninitiated');
-    const [autoScroll, setAutoScroll] = React.useState<boolean>(true);
-    const [lastUpdate, setLastUpdate] = React.useState<string | null>('');
-
-    React.useEffect(() => {
-        if (props.StatsURL == null || props.StatsURL.length == 0) return;
-
-        setStatus('loading');
-        let statHandle: JQuery.jqXHR<string>;
-        const intervalHandle = setInterval(() => {
-            if (statHandle != null && statHandle?.abort != null) statHandle.abort();
-
-            statHandle = $.ajax({
-                type: "GET",
-                url: props.StatsURL,
-                dataType: 'text',
-                cache: false,
-                async: true
-            }).done((stat: string) => {
-                setStatus('idle');
-                // Regex remove wrapping quotes, carriage return, and format \\ to \
-                setStatInfo(stat.replace(/^\"+|\"+$/g, '').replace(/\\r\\n/g, '\n').replace(/\\\\/g, '\\'));
-            }).fail((_a, _b, e) => {
-                console.error(e);
-                setStatus('error');
-            });
-        }, 5000);
-
-        return () => {
-            clearInterval(intervalHandle);
-            if (statHandle != null && statHandle.abort != null) statHandle.abort();
-        };
-    }, [props.ApplicationName]);
 
     return (
         <Modal
             Show={props.ApplicationName !== '' ? true : false}
             ShowCancel={false} ShowConfirm={false} ShowX={true} Size={'xlg'}
-            CallBack={() => { props.Close(); setStatus('uninitiated'); setStatInfo('') }}
-            Title={'Statistics - ' + props.ApplicationName}
+            CallBack={() => { props.Close(); /**setStatus('uninitiated'); setStatInfo('')*/ }}
+            Title={'Details - ' + props.ApplicationName}
         >
-            {(props.StatsURL ?? "") === "" ? // only set if it's openXDA. certainly a better way to do this dear natalie. move XDA health into another component.
-                <div className="col-12">
-                    <NodeHealth
-                        ApplicationName={props.ApplicationName}
-                        ApplicationType={props.ApplicationType}
-                        Properties={props.Properties}
-                    />
-                </div> :
-                <div className="row">
-                    <fieldset className="border col-6" style={{ padding: '10px', height: '100%' }}>
-                        <legend className="w-auto" style={{ fontSize: 'large' }}>XDA Stats:</legend>
-                        {
-                            status === "error" ?
-
-                                <div className={`col-12 d-flex alert-danger`}>
-                                    <span className={"my-3"}>
-                                        <ReactIcons.CircledX Color="var(--danger)" />
-                                    </span>
-                                    <h5
-                                        className={"m-3"}
-                                    >
-                                       Failed to get openXDA stats.
-                                    </h5>
-                                </div> :
-                                status === "loading" ?
-                                    <ReactIcons.SpiningIcon /> :
-                                    <div className="w-100 h-100">
-                                        <pre style={statStyle}>
-                                            {statInfo}
-                                        </pre>
-                                    </div>
-                        }
-                    </fieldset>
-                    <div className="col-6">
-                        <NodeHealth
-                            ApplicationName={props.ApplicationName}
-                            ApplicationType={props.ApplicationType}
-                            Properties={props.Properties}
-                        />
-                    </div>
-                </div>
-            }
-
+            <div className="col-12">
+                <NodeHealth
+                    ApplicationName={props.ApplicationName}
+                    ApplicationType={props.ApplicationType}
+                    Properties={props.Properties}
+                />
+            </div>
         </Modal>);
 }
